refactor: split launch wizard quick start loading
